+++ settings.gradle
@@ -60,7 +60,6 @@ include ':simplified-reader-api'
 include ':simplified-reader-bookmarks'
 include ':simplified-reader-bookmarks-api'
 include ':simplified-reports'
-include ':simplified-splash'
 include ':simplified-stack'
 include ':simplified-taskrecorder-api'
 include ':simplified-tenprint'
@@ -70,5 +69,6 @@ include ':simplified-tests-sandbox'
 include ':simplified-tests-strings'
 include ':simplified-threads'
 include ':simplified-ui-branding'
+include ':simplified-ui-splash'
 include ':simplified-ui-theme'
