+++ simplified-app-shared/build.gradle
@@ -72,12 +72,12 @@ dependencies {
   api project(":simplified-reader-bookmarks")
   api project(":simplified-reader-bookmarks-api")
   api project(":simplified-reports")
-  api project(":simplified-splash")
   api project(":simplified-stack")
   api project(":simplified-tenprint")
-  api project(":simplified-ui-theme")
   api project(":simplified-threads")
   api project(":simplified-ui-branding")
+  api project(":simplified-ui-splash")
+  api project(":simplified-ui-theme")
 
   annotationProcessor libraries.googleAutoValue
