+++ simplified-app-shared/src/main/java/org/nypl/simplified/app/splash/SplashActivity.kt
@@ -28,9 +28,9 @@ import org.nypl.simplified.profiles.api.ProfilesDatabaseType
 import org.nypl.simplified.profiles.api.ProfilesDatabaseType.AnonymousProfileEnabled.ANONYMOUS_PROFILE_DISABLED
 import org.nypl.simplified.profiles.api.ProfilesDatabaseType.AnonymousProfileEnabled.ANONYMOUS_PROFILE_ENABLED
 import org.nypl.simplified.profiles.controller.api.ProfilesControllerType
-import org.nypl.simplified.splash.SplashFragment
-import org.nypl.simplified.splash.SplashListenerType
-import org.nypl.simplified.splash.SplashParameters
+import org.nypl.simplified.ui.splash.SplashFragment
+import org.nypl.simplified.ui.splash.SplashListenerType
+import org.nypl.simplified.ui.splash.SplashParameters
 import org.nypl.simplified.taskrecorder.api.TaskResult
 import org.nypl.simplified.ui.theme.ThemeControl
 import org.nypl.simplified.threads.NamedThreadPools
