+++ simplified-app-vanilla/version.properties
@@ -1,4 +1,4 @@
 #
-#Thu Oct 10 09:58:21 UTC 2019
+#Thu Oct 10 10:35:10 UTC 2019
 versionName=3.0.14
-versionCode=4165
+versionCode=4170
