--- simplified-splash/gradle.properties
@@ -1,4 +1,0 @@
-POM_ARTIFACT_ID=org.librarysimplified.splash
-POM_DESCRIPTION=Library Simplified (Splash screen)
-POM_NAME=org.librarysimplified.splash
-POM_PACKAGING=aar
