+++ simplified-ui-splash/build.gradle
