+++ simplified-ui-splash/gradle.properties
@@ -1,0 +1,4 @@
+POM_ARTIFACT_ID=org.librarysimplified.ui.splash
+POM_DESCRIPTION=Library Simplified (Splash screen)
+POM_NAME=org.librarysimplified.ui.splash
+POM_PACKAGING=aar
