+++ simplified-ui-splash/src/main/java/org/nypl/simplified/ui/splash/SplashColorMatrix.java
@@ -1,4 +1,4 @@
-package org.nypl.simplified.splash;
+package org.nypl.simplified.ui.splash;
 
 import android.graphics.Color;
 import android.graphics.ColorMatrix;
