+++ simplified-ui-splash/src/main/java/org/nypl/simplified/ui/splash/SplashListenerType.kt
@@ -1,4 +1,4 @@
-package org.nypl.simplified.splash
+package org.nypl.simplified.ui.splash
 
 import com.google.common.util.concurrent.ListenableFuture
 import com.google.common.util.concurrent.ListeningExecutorService
