+++ simplified-ui-splash/src/main/res/anim/zoom_fade.xml
