+++ simplified-ui-splash/src/main/res/drawable/border.xml
