+++ simplified-ui-splash/src/main/res/drawable/migration_icon.png
