+++ simplified-ui-splash/src/main/res/layout/splash_eula.xml
