+++ simplified-ui-splash/src/main/res/layout/splash_migration_running.xml
