+++ simplified-ui-splash/src/main/res/layout/splash_stack.xml
