+++ simplified-ui-splash/src/main/res/values-de/strings.xml
