+++ simplified-ui-splash/src/main/res/values-es/strings.xml
