+++ simplified-ui-splash/src/main/res/values-it/strings.xml
