+++ simplified-ui-splash/src/main/res/values/strings.xml
